+++ changelog.md
@@ -1,5 +1,9 @@
+4.0.0-beta13 - November 2, 2016
+- Correctly expanding types within uriParameters
+- Updated datatype-expansion and raml-1-parser
+
 4.0.0-beta12 - November 1, 2016
-- Handling `array` types with `items` by expanding the items into a type object 
+- Handling `array` types with `items` by expanding the items into a type object
 
 4.0.0-beta11 - November 1, 2016
 - Updated datatype-expansion to 0.0.14